Get composer.json relative to current file, not based on the current working directory
Using the current working directory to identify the vendor directory is not safe as artisan may be executed from a directory other than the Laravel root.  Instead, use the __DIR__ constant which identifies the directory of the file we’re currently in, and then work out where the composer.json should be from there.